+++ src/Common/StyleConstants.ts
@@ -1,0 +1,2 @@
+export let StyleConstants = require("less-vars-loader!../../less/Common/Constants.less");
+
